leanvibe-cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "leanvibe-cli"
version = "0.1.0"
description = "LeanVibe CLI - Terminal interface for enterprise codebase analysis"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "LeanVibe Team" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "websockets>=12.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "httpx>=0.25.2",
    "asyncio-throttle>=1.0.2",
    "typer>=0.9.0", # For future enhancement
    "pydantic>=2.5.0",
    "plyer>=2.1.0", # Cross-platform notifications
    "win10toast>=0.9; sys_platform == 'win32'", # Windows notifications
    "qrcode>=8.2", # QR code generation
    "websocket-client>=1.8.0",
    "requests>=2.32.4",
    "markdown-it-py>=3.0.0",
    "pygments>=2.19.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "psutil>=5.9.0",
    "aiohttp>=3.9.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/leanvibe-ai/leanvibe-cli"
Repository = "https://github.com/leanvibe-ai/leanvibe-cli"

[project.scripts]
leanvibe = "leanvibe_cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["leanvibe_cli"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
